requirements.txt
# Slack Bolt framework - official framework for building Slack apps with Python
slack-bolt>=1.18.0

# Python dotenv for environment variable management
python-dotenv>=1.0.0

# PostgreSQL adapter for Python
psycopg2-binary>=2.9.9

# Optional: For async support (uncomment if needed)
# aiohttp>=3.9.0

# OpenAI
openai

# Schedule https://schedule.readthedocs.io/en/stable/index.html
schedule
pytz

# Google
google-api-python-client
oauth2client
